separate out display code further, reword comments on InstructionTextSink to be ... stern
