+++ src/lib.rs
@@ -139,6 +139,8 @@ pub mod real_mode;
 pub use real_mode::Arch as x86_16;
 
 mod safer_unchecked;
+#[cfg(feature = "fmt")]
+mod display;
 
 const MEM_SIZE_STRINGS: [&'static str; 65] = [
     "BUG",
